refactor(schema): extract complex schema types to dedicated files
