+++ src/validators/schema/and.ts
@@ -1,9 +1,12 @@
 import type { TypeGuard } from '../../TypeGuards/types/index.ts'
 import type { Merge } from '../../types/index.ts'
 import type { Custom } from '../rules/types/index.ts'
-import type { StandardSchemaV1 } from '../standard-schema/types.ts'
 import type { V3 } from './types/index.ts'
-import type { FluentSchema } from './types/FluentSchema.ts'
+import type {
+    IntersectionSchema,
+    IntersectionSchemaEntry,
+    GetIntersectionEntryTypes,
+} from './types/IntersectionSchema.ts'
 
 import { getMessage } from '../../TypeGuards/helpers/getMessage.ts'
 import { useCustomRules } from '../rules/helpers/useCustomRules.ts'
@@ -17,17 +20,6 @@ import { setRuleMessage } from './helpers/setRuleMessage.ts'
 import { setStructMetadata } from './helpers/setStructMetadata.ts'
 import { validateCustomRules } from './helpers/validateCustomRules.ts'
 import { toStandardSchema } from '../standard-schema/toStandardSchema.ts'
-
-type IntersectionSchemaEntry<T = any> = TypeGuard<T> | StandardSchemaV1<T, T>
-
-type GetIntersectionEntryType<T> =
-    T extends TypeGuard<infer U> ? U : T extends StandardSchemaV1<infer U, any> ? U : never
-
-type GetIntersectionEntryTypes<T extends any[]> = T extends []
-    ? []
-    : T extends [infer U, ...infer V]
-      ? [GetIntersectionEntryType<U>, ...GetIntersectionEntryTypes<V>]
-      : any[]
 
 function _fn<T1, T2>(
     guard1: IntersectionSchemaEntry<T1>,
@@ -77,22 +69,6 @@ type OptionalizedAnd = {
 }
 
 export const _and = optionalizeOverloadFactory(_fn).optionalize<OptionalizedAnd>()
-
-type IntersectionSchema = CallableFunction & {
-    <T1, T2>(
-        guard1: IntersectionSchemaEntry<T1>,
-        guard2: IntersectionSchemaEntry<T2>
-    ): FluentSchema<T1 & T2>
-    <
-        TEntries extends [
-            IntersectionSchemaEntry<any>,
-            IntersectionSchemaEntry<any>,
-            ...IntersectionSchemaEntry[],
-        ],
-    >(
-        ...guards: TEntries
-    ): FluentSchema<V3.TIntersection<GetIntersectionEntryTypes<TEntries>>>
-}
 
 export const and: IntersectionSchema = ((
     guard1: IntersectionSchemaEntry<any>,
